Finished the basic Champion Db, and started the webpage, with components and routing
The champion Db now is able to take data but I need to rework the classes to take data and edit them. The webpage is now serveable with a component for the header and footer with a middle for Champions. Basic routing has been set up for the component, and can now request data from my API for champions using the client. The data isnt being displayed yet but it can be seen with console.log.

+++ CaLOLateAPI/CaLOLateAPI/Models/Champion.cs
@@ -17,12 +17,12 @@ namespace CaLOLateAPI.Models
         public int Level { get; set; }
         public int HeathFlat { get; set; }
         public double HeathScaling { get; set; }
-        public int HealthRegFlat { get; set; }
+        public double HealthRegFlat { get; set; }
         public double HealthRegScaling { get; set; }
         //TODO:Mana/Mana regen/HpCost/Rage
         public int AdFlat { get; set;}
         public double AdScaling { get; set; }
-        public int AttackSpeedFlat { get; set; }
+        public double AttackSpeedFlat { get; set; }
         public double AttackSpeedScale { get; set; }
         public int ArmourFlat { get; set; }
         public double ArmourScaling { get; set; }
